+++ apps/laboratory/tests/multichain/multichain-wagmi-solana-siwe.spec.ts
@@ -1,9 +1,9 @@
 import { type BrowserContext, test } from '@playwright/test'
 
+import { WalletPage, WalletValidator } from '@reown/appkit-testing'
+
 import { ModalPage } from '../shared/pages/ModalPage'
-import { WalletPage } from '../shared/pages/WalletPage'
 import { ModalValidator } from '../shared/validators/ModalValidator'
-import { WalletValidator } from '../shared/validators/WalletValidator'
 
 /* eslint-disable init-declarations */
 let modalPage: ModalPage