+++ packages/adapters/bitcoin/src/connectors/LeatherConnector.ts
@@ -5,9 +5,9 @@ import type {
   SignPsbtResponseBody
 } from '@leather.io/rpc'
 
+import type { BitcoinConnector } from '@reown/appkit-utils/bitcoin'
 import { bitcoin, bitcoinTestnet } from '@reown/appkit/networks'
 
-import type { BitcoinConnector } from '../index.js'
 import { SatsConnectConnector } from './SatsConnectConnector.js'
 
 export class LeatherConnector extends SatsConnectConnector {