+++ packages/adapters/bitcoin/src/connectors/WalletStandardConnector.ts
@@ -5,10 +5,10 @@ import type { Wallet, WalletWithFeatures } from '@wallet-standard/base'
 import type { CaipNetwork } from '@reown/appkit-common'
 import type { Provider, RequestArguments } from '@reown/appkit-controllers'
 import { PresetsUtil } from '@reown/appkit-utils'
+import type { BitcoinConnector } from '@reown/appkit-utils/bitcoin'
 import { bitcoin, bitcoinTestnet } from '@reown/appkit/networks'
 
 import { MethodNotSupportedError } from '../errors/MethodNotSupportedError.js'
-import type { BitcoinConnector } from '../utils/BitcoinConnector.js'
 import { AddressPurpose } from '../utils/BitcoinConnector.js'
 import { ProviderEventEmitter } from '../utils/ProviderEventEmitter.js'
 import type { BitcoinFeatures } from '../utils/wallet-standard/WalletFeatures.js'
@@ -20,8 +20,6 @@ export class WalletStandardConnector extends ProviderEventEmitter implements Bit
   readonly provider: Provider
   readonly wallet: Wallet
   private requestedChains: CaipNetwork[] = []
-
-  private walletUnsubscribes: (() => void)[] = []
 
   constructor({ wallet, requestedChains }: WalletStandardConnector.ConstructorParams) {
     super()
@@ -67,7 +65,6 @@ export class WalletStandardConnector extends ProviderEventEmitter implements Bit
   async connect() {
     const connectFeature = this.getWalletFeature('bitcoin:connect')
 
-    this.bindEvents()
     const response = await connectFeature.connect({ purposes: ['payment', 'ordinals'] })
 
     const account = response.accounts[0]
@@ -178,8 +175,6 @@ export class WalletStandardConnector extends ProviderEventEmitter implements Bit
   }
 
   async disconnect() {
-    this.unbindEvents()
-
     return Promise.resolve()
   }
 
@@ -195,38 +190,6 @@ export class WalletStandardConnector extends ProviderEventEmitter implements Bit
     return this.wallet.features[feature] as WalletWithFeatures<
       Record<Name, BitcoinFeatures[Name]>
     >['features'][Name]
-  }
-
-  private bindEvents() {
-    this.unbindEvents()
-
-    try {
-      const feature = this.getWalletFeature('standard:events')
-
-      this.walletUnsubscribes.push(
-        feature.on('change', data => {
-          if ('accounts' in data && data.accounts) {
-            if (data.accounts.length === 0) {
-              this.emit('disconnect')
-            } else {
-              this.emit(
-                'accountsChanged',
-                data.accounts.map(acc => acc.address)
-              )
-            }
-          }
-        })
-      )
-    } catch {
-      console.warn(
-        `WalletStandardConnector:bindEvents - wallet provider "${this.name}" does not support events`
-      )
-    }
-  }
-
-  private unbindEvents() {
-    this.walletUnsubscribes.forEach(unsubscribe => unsubscribe())
-    this.walletUnsubscribes = []
   }
 
   public static watchWallets({